Add custom Sysctl args and methods SysctlUint64, SysctlRaw
Add support for optional sysctl arguments which is required to support sysctls that require more than the mib identifer args as returned from nametomib such as kern.proc.pid.

Add SysctlUint64 which allows sysctls that return 64 bit ints to be queried.

Add SysctlRaw which allows sysctls that return structs or other unsupported types to be queried.