Add support for PRE_SHUTDOWN service control event
Changes the handler code to use RegisterServiceCtrlHandlerExW
instead of the older RegisterServiceCtrlHandlerW function.
This allows handlers to then receive the new extended control
signals such as PRE_SHUTDOWN which allows for blocking of the
shutdown process for long cleanups on versions of Windows
above Win2k3.

The assembly code for amd64 and 386 has been modified to
support the different call signature of the
RegisterServiceCtrlHandlerExW procedure and the handler
modified to comply with the extended callback/handler call
signature required by the new procedure.

Change-Id: Ia46e8cf9210e5a4a59bc04f33e454479664fa05c